Move LifecycleWallpaperService to util package
